Split profile persistence module
